--- drivers/SmartThings/zwave-switch/profiles/multichannel-switch-binary.yml
@@ -1,12 +1,0 @@
-name: multichannel-switch-binary
-components:
-- id: main
-  capabilities:
-  - id: switch
-    version: 1
-  - id: refresh
-    version: 1
-  - id: zwMultichannel
-    version: 1
-  categories:
-  - name: Switch
